Fix failing second call of GlobalInit() when first call failed.
LoadManagerBase<T>.GlobalInit() will throw an exception when it failed
to load a native library path. However it failed to reset Lib reference
back to null. It prevented reloading GlobalInit() with right path,
causing a soft-lock.

This commit fixes this by properly resetting Lib ref to null when
LoadLibrary() or hooks throwed an exception.